[js] Move the Builder class to the main module and remove the builder module. The builder module was originally added to avoid a circular dependency that no longer exists.
